+++ package.json
@@ -35,7 +35,8 @@
         "react-dom": "^17.0.1",
         "react-router-dom": "^5.2.0",
         "react-scripts": "4.0.1",
-        "web-vitals": "^0.2.4"
+        "web-vitals": "^0.2.4",
+        "whoami": "0.0.3"
     },
     "scripts": {
         "start": "react-scripts start",
